requirements.txt
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.4.0.2
certifi==2017.7.27.1
chardet==3.0.4
Cheetah==2.4.4
click==6.7
cloud-init==0.7.6
configobj==4.7.2
decorator==3.4.0
Flask==1.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.6
iniparse==0.4
itsdangerous==0.24
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.12
Markdown==2.6.9
MarkupSafe==1.0
numpy==1.15.1
oauth==1.0.1
perf==0.1
Pillow==5.2.0
prettytable==0.7.2
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
pyparsing==2.2.0
python-linux-procfs==0.4.9
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.12
qrcode==6.0
requests==2.18.4
schedutils==0.4
six==1.11.0
slip==0.4.0
slip.dbus==0.4.0
urlgrabber==3.10
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.2.1
yum-metadata-parser==1.1.4
